+++ server/src/config/controllers.js
@@ -42,9 +42,7 @@ export const initializeControllers = (services) => {
 	controllers.logController = new LogController(services.logger);
 	controllers.statusPageController = new StatusPageController(services.db);
 	controllers.notificationController = new NotificationController(services.notificationService, services.db);
-	controllers.diagnosticController = new DiagnosticController(commonDependencies, {
-		diagnosticService: services.diagnosticService,
-	});
+	controllers.diagnosticController = new DiagnosticController(services.diagnosticService);
 
 	controllers.incidentController = new IncidentController(services.incidentService);
